server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
colorama==0.4.6
fastapi==0.127.0
fire==0.7.1
fonttools==4.61.1
h11==0.16.0
idna==3.11
lxml==6.0.2
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pdf2docx==0.5.8
pdf2image==1.17.0
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.26.7
PyPDF2==3.0.1
pytesseract==0.3.13
python-docx==1.2.0
python-multipart==0.0.21
starlette==0.50.0
termcolor==3.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
